double to Decimal.

+++ lib/screens/home/view_more/hidden_proposal_creation/hidden_proposal_creation_screen.dart
@@ -173,13 +173,13 @@ class _HiddenProposalCreationScreenState
                           children: [
                             PwText(
                               strings.hashAmount(
-                                  details.initialDeposit.toInt().toString()),
+                                  details.initialDeposit.toString()),
                               style: PwTextStyle.footnote,
                             )
                           ],
                         ),
                         DepositSlider(
-                          max: details.hashAmount,
+                          max: details.hashAmount.toDouble(),
                           thumbColor: Theme.of(context).colorScheme.primary550,
                           onChanged: (changed) =>
                               _bloc.updateInitialDeposit(changed),
